Remove some string getter from enum + Master excel now uses id as key value + Change script to use id

+++ Assets/Scripts/_Runtime/Game/Item/ItemTypeSet.cs
@@ -17,7 +17,7 @@ namespace Personal.Item
 		protected override void Initialize()
 		{
 			if (itemType == default) return;
-			Entity = MasterDataManager.Instance.Item.Get(itemType);
+			Entity = MasterDataManager.Instance.Item.Get((int)itemType);
 		}
 
 		void IItem.Use()
